profiler/pyproject.toml
[project]
name = "profilier"
version = "0.1.0"
description = ""
authors = [
    {name = "Tylores",email = "tylor.slay@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "tsgm",
    "jupyter",
    "seaborn",
    "matplotlib",
    "numpy",
    "pandas",
    'OpenDSSDirect.py >= 0.9.4',
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
